Улучшил панель, объекты теперь могут умирать и заболевать заново

+++ src/tools/Instruments.java
@@ -2,9 +2,9 @@ package tools;
 import java.util.Random;
 
 public class Instruments {
-    public static int random_number(int start, int end) {
+    public static double random_number(double start, double end) {
         Random random_num = new Random();
-        return random_num.nextInt(start, end);
+        return random_num.nextDouble(start, end);
     }
 
 }
